requirements.txt
Django==2.2.4
pycryptodome==3.11.0
